+++ src/LinearMaps.jl
@@ -1,25 +1,25 @@
 module LinearMaps
 
-export LinearMap
+export LinearMap, FunctionMap, FillMap, InverseMap
 export ⊗, squarekron, kronsum, ⊕, sumkronsum, khatrirao, facesplitting
-export FillMap
-export InverseMap
 
 using LinearAlgebra
+using LinearAlgebra: AbstractQ
 import LinearAlgebra: mul!
 using SparseArrays
 
 import Statistics: mean
 
-using ChainRulesCore: unthunk, NoTangent, @thunk
+using ChainRulesCore: unthunk, NoTangent, @thunk, @not_implemented
 import ChainRulesCore: rrule
 
 using Base: require_one_based_indexing
 
 abstract type LinearMap{T} end
 
-const MapOrVecOrMat{T} = Union{LinearMap{T}, AbstractVecOrMat{T}}
-const MapOrMatrix{T} = Union{LinearMap{T}, AbstractMatrix{T}}
+const AbstractVecOrMatOrQ{T} = Union{AbstractVecOrMat{T}, AbstractQ{T}}
+const MapOrVecOrMat{T} = Union{LinearMap{T}, AbstractVecOrMatOrQ{T}}
+const MapOrMatrix{T} = Union{LinearMap{T}, AbstractMatrix{T}, AbstractQ{T}}
 const TransposeAbsVecOrMat{T} = Transpose{T,<:AbstractVecOrMat}
 const RealOrComplex = Union{Real, Complex}
 
@@ -31,7 +31,7 @@ Base.eltype(::LinearMap{T}) where {T} = T
 
 # conversion to LinearMap
 Base.convert(::Type{LinearMap}, A::LinearMap) = A
-Base.convert(::Type{LinearMap}, A::AbstractVecOrMat) = LinearMap(A)
+Base.convert(::Type{LinearMap}, A::AbstractVecOrMatOrQ) = LinearMap(A)
 
 convert_to_lmaps() = ()
 convert_to_lmaps(A) = (convert(LinearMap, A),)
@@ -40,15 +40,22 @@ convert_to_lmaps(A) = (convert(LinearMap, A),)
 
 abstract type MulStyle end
 
-struct FiveArg <: MulStyle end
-struct ThreeArg <: MulStyle end
+struct FiveArg <: MulStyle end # types admit in-place multiplication and addition
+struct ThreeArg <: MulStyle end # types "only" admit in-place multiplication
+struct TwoArg <: MulStyle end # types "only" admit out-of-place multiplication
 
 MulStyle(::FiveArg, ::FiveArg) = FiveArg()
-MulStyle(::ThreeArg, ::FiveArg) = ThreeArg()
 MulStyle(::FiveArg, ::ThreeArg) = ThreeArg()
+MulStyle(::FiveArg, ::TwoArg) = TwoArg()
+MulStyle(::ThreeArg, ::FiveArg) = ThreeArg()
 MulStyle(::ThreeArg, ::ThreeArg) = ThreeArg()
+MulStyle(::ThreeArg, ::TwoArg) = ThreeArg()
+MulStyle(::TwoArg, ::FiveArg) = TwoArg()
+MulStyle(::TwoArg, ::ThreeArg) = ThreeArg()
+MulStyle(::TwoArg, ::TwoArg) = TwoArg()
 MulStyle(::LinearMap) = ThreeArg() # default
 MulStyle(::AbstractVecOrMat) = FiveArg()
+MulStyle(::AbstractQ) = ThreeArg()
 MulStyle(A::LinearMap, As::LinearMap...) = MulStyle(MulStyle(A), MulStyle(As...))
 
 Base.isreal(A::LinearMap) = eltype(A) <: Real
@@ -112,10 +119,6 @@ _combine(As::LinearMapVector, Bs::LinearMapVector) = Base.vect(As..., Bs...)
 
 Compute the action of the linear map `A` on the vector `x`.
 
-!!! compat "Julia 1.3"
-    In Julia versions v1.3 and above, objects `L` of any subtype of `LinearMap`
-    are callable in the sense that `L(x) = L*x` for `x::AbstractVector`.
-
 ## Examples
 ```jldoctest; setup=(using LinearAlgebra, LinearMaps)
 julia> A=LinearMap([1.0 2.0; 3.0 4.0]); x=[1.0, 1.0];
@@ -135,7 +138,7 @@ function Base.:(*)(A::LinearMap, x::AbstractVector)
     check_dim_mul(A, x)
     T = typeof(oneunit(eltype(A)) * oneunit(eltype(x)))
     y = similar(x, T, axes(A)[1])
-    return mul!(y, A, x)
+    return @inbounds mul!(y, A, x)
 end
 
 (L::LinearMap)(x::AbstractVector) = L*x
@@ -165,8 +168,8 @@ julia> Y
  7.0  7.0
 ```
 """
-function mul!(y::AbstractVecOrMat, A::LinearMap, x::AbstractVector)
-    check_dim_mul(y, A, x)
+@inline function mul!(y::AbstractVecOrMat, A::LinearMap, x::AbstractVector)
+    @boundscheck check_dim_mul(y, A, x)
     return _unsafe_mul!(y, A, x)
 end
 # the following is of interest in, e.g., subspace-iteration methods
@@ -193,7 +196,7 @@ julia> mul!(Y, A, b)
 ```
 """
 function mul!(y::AbstractVecOrMat, A::LinearMap, s::Number)
-    size(y) == size(A) ||     
+    size(y) == size(A) ||
         throw(
             DimensionMismatch("y has size $(size(y)), A has size $(size(A))."))
     return _unsafe_mul!(y, A, s)
@@ -256,7 +259,7 @@ julia> mul!(Y, A, b, 2, 1)
 ```
 """
 function mul!(y::AbstractMatrix, A::LinearMap, s::Number, α::Number, β::Number)
-    size(y) == size(A) ||     
+    size(y) == size(A) ||
         throw(
             DimensionMismatch("y has size $(size(y)), A has size $(size(A))."))
     return _unsafe_mul!(y, A, s, α, β)
@@ -335,11 +338,11 @@ end
 include("transpose.jl") # transposing linear maps
 include("wrappedmap.jl") # wrap a matrix of linear map in a new type, thereby allowing to alter its properties
 include("left.jl") # left multiplication by a matrix/transpose or adjoint vector
+include("functionmap.jl") # using a function as linear map
 include("uniformscalingmap.jl") # the uniform scaling map, to be able to make linear combinations of LinearMap objects and multiples of I
 include("linearcombination.jl") # defining linear combinations of linear maps
 include("scaledmap.jl") # multiply by a (real or complex) scalar
 include("composition.jl") # composition of linear maps
-include("functionmap.jl") # using a function as linear map
 include("blockmap.jl") # block linear maps
 include("kronecker.jl") # Kronecker product of linear maps
 include("khatrirao.jl") # Khatri-Rao and face-splitting products
@@ -353,7 +356,7 @@ include("chainrules.jl") # AD rules through ChainRulesCore
 
 """
     LinearMap(A::LinearMap; kwargs...)::WrappedMap
-    LinearMap(A::AbstractVecOrMat; kwargs...)::WrappedMap
+    LinearMap(A::AbstractVecOrMatOrQ; kwargs...)::WrappedMap
     LinearMap(J::UniformScaling, M::Int)::UniformScalingMap
     LinearMap{T=Float64}(f, [fc,], M::Int, N::Int = M; kwargs...)::FunctionMap
     LinearMap(A::MapOrVecOrMat, dims::Dims{2}, index::NTuple{2, AbstractVector{Int}})::EmbeddedMap
@@ -361,26 +364,28 @@ include("chainrules.jl") # AD rules through ChainRulesCore
 
 Construct a linear map object, either
 
-1. from an existing `LinearMap` or `AbstractVecOrMat` `A`, with the purpose of redefining
-  its properties via the keyword arguments `kwargs`;
+1. from an existing `LinearMap` or `AbstractVecOrMat`/`AbstractQ` `A`, with the purpose of
+  redefining its properties via the keyword arguments `kwargs`, see below;
 2. a `UniformScaling` object `J` with specified (square) dimension `M`;
 3. from a function or callable object `f`;
-4. from an existing `LinearMap` or `AbstractVecOrMat` `A`, embedded in a larger zero map.
+4. from an existing `LinearMap` or `AbstractVecOrMat`/`AbstractQ` `A`, embedded in a larger
+   zero map.
 
 In the case of item 3, one also needs to specify the size of the equivalent matrix
-representation `(M, N)`, i.e., for functions `f` acting
-on length `N` vectors and producing length `M` vectors (with default value `N=M`).
-Preferably, also the `eltype` `T` of the corresponding matrix representation needs to be
-specified, i.e., whether the action of `f` on a vector will be similar to, e.g., multiplying
-by numbers of type `T`. If not specified, the devault value `T=Float64` will be assumed.
-Optionally, a corresponding function `fc` can be specified that implements the adjoint
-(=transpose in the real case) of `f`.
+representation `(M, N)`, i.e., for functions `f` acting on length `N` vectors and producing
+length `M` vectors (with default value `N=M`). Preferably, also the `eltype` `T` of the
+corresponding matrix representation needs to be specified, i.e., whether the action of `f`
+on a vector will be similar to, e.g., multiplying by numbers of type `T`. If not specified,
+the devault value `T=Float64` will be assumed. Optionally, a corresponding function `fc`
+can be specified that implements the adjoint (or transpose in the real case) of `f`.
 
-The keyword arguments and their default values for the function-based constructor are:
-*   `issymmetric::Bool = false` : whether `A` or `f` act as a symmetric matrix
-*   `ishermitian::Bool = issymmetric & T<:Real` : whether `A` or `f` act as a Hermitian
-    matrix
-*   `isposdef::Bool = false` : whether `A` or `f` act as a positive definite matrix.
+The keyword arguments and their default values are:
+
+* `issymmetric::Bool = false` : whether `A` or `f` act as a symmetric matrix
+* `ishermitian::Bool = issymmetric & T<:Real` : whether `A` or `f` act as a Hermitian
+  matrix
+* `isposdef::Bool = false` : whether `A` or `f` act as a positive definite matrix.
+
 For existing linear maps or matrices `A`, the default values will be taken by calling
 internal functions `_issymmetric`, `_ishermitian` and `_isposdef` on the existing object `A`.
 These in turn dispatch to (overloads of) `LinearAlgebra`'s `issymmetric`, `ishermitian`,
@@ -389,11 +394,11 @@ known at compile time as for certain structured matrices, but return `false` for
 `AbstractMatrix` types.
 
 For the function-based constructor, there is one more keyword argument:
-*   `ismutating::Bool` : flags whether the function acts as a mutating matrix multiplication
-    `f(y,x)` where the result vector `y` is the first argument (in case of `true`),
-    or as a normal matrix multiplication that is called as `y=f(x)` (in case of `false`).
-    The default value is guessed by looking at the number of arguments of the first
-    occurrence of `f` in the method table.
+* `ismutating::Bool` : flags whether the function acts as a mutating matrix multiplication
+  `f(y,x)` where the result vector `y` is the first argument (in case of `true`),
+  or as a normal matrix multiplication that is called as `y=f(x)` (in case of `false`).
+  The default value is guessed by looking at the number of arguments of the first
+  occurrence of `f` in the method table.
 
 For the `EmbeddedMap` constructors, `dims` specifies the total dimensions of the map. The
 `index` argument specifies two collections of indices `inds1` and `inds2`, such that for